Fix to prevent client from uploading private files (if only we can see them). We can remove this once we add e2e encryption.

+++ syftbox/client/plugins/sync.py
@@ -116,10 +116,19 @@ def filter_changes(
             if (
                 user_email in perm_file_at_path.write
                 or "GLOBAL" in perm_file_at_path.write
-            ):
+            ) and perm_file_at_path.read != [user_email]: # skip upload if only we have read perms.
+                
                 valid_changes.append(change)
                 valid_change_files.append(change.sub_path)
                 continue
+            elif perm_file_at_path.read == [user_email]:
+                
+                if change.internal_path[-10:] == "_.syftperm":
+                    # include changes for syft_perm file even if only we have read perms.
+                    valid_changes.append(change)
+                    valid_change_files.append(change.sub_path)
+                    continue
+
         invalid_changes.append(change)
     return valid_changes, valid_change_files, invalid_changes
 
@@ -300,7 +309,7 @@ def sync_up(client_config):
 
         synced_dir_state = prune_invalid_changes(new_dir_state, changed_files)
 
-        # combine successfulc hanges qwith old dir state
+        # combine successful changes qwith old dir state
         combined_tree = old_dir_state.tree
         combined_tree.update(synced_dir_state.tree)
         synced_dir_state.tree = combined_tree
@@ -337,6 +346,9 @@ def sync_down(client_config) -> int:
         # get the top level perm file
 
         dir_filename = f"{change_log_folder}/{datasite}.json"
+        
+        datasite_path = os.path.join(client_config.sync_folder, datasite)
+        perm_tree = PermissionTree.from_path(datasite_path)
 
         # get the new dir state
         new_dir_state = hash_dir(client_config.sync_folder, datasite, IGNORE_FOLDERS)
@@ -372,6 +384,15 @@ def sync_down(client_config) -> int:
         for change in changes:
             change.sync_folder = client_config.sync_folder
             if change.kind_delete:
+                perm_file_at_path = perm_tree.permission_for_path(change.sub_path)
+                if perm_file_at_path.read != [client_config.email]:
+                    # the only reason we're wanting to delete this is because
+                    # we skipped sending it to the server. and the raeson we 
+                    # skipped sending it to the server is because it's a file that
+                    # only we can read, and so we don't want to share it with the
+                    # server... who we might not trust. Aka... this is a private file/folder
+                    # only for us.
+                    continue
                 result = change.delete()
                 if result:
                     deleted_files.append(change.internal_path)
@@ -384,7 +405,6 @@ def sync_down(client_config) -> int:
         synced_dir_state.tree = combined_tree
 
         synced_dir_state = delete_files(new_dir_state, deleted_files)
-        change_text = ""
 
         change_text = ""
         if len(changed_files):
@@ -393,8 +413,10 @@ def sync_down(client_config) -> int:
         if len(deleted_files):
             change_text += f"❌ Syncing Down {len(deleted_files)} Deletes\n"
             change_text += ascii_for_change(deleted_files)
-
-        print(change_text)
+        
+        if(len(change_text) > 0):
+            print(change_text)
+        
         synced_dir_state.save(dir_filename)
         n_changes += len(changed_files) + len(deleted_files)
